refactor(motor sound): one clamp, one FIR stream, and the ten missing voices
accel_voices arrived with its own copy of motor_sound's _clamp01 and with
Stream, a tail-carrying FIR that motor_synth had already hand-rolled three
times over. Both now live in motor_synth beside Rotor and bandpass_kernel,
which the voices already import for exactly this reason. The whine renders
bit-identical across a drive cycle -- same maths, one copy of it.

The module docstring stopped at 24 voices while the registry grew to 34.
The ten it never mentioned -- the impossible ones, the ones borrowing
speed from another domain, and the two alive ones -- are listed now.
